Website for Educational Institutes
 A web based system developed using React JS, Spring 6, Maven 3, Spring boot 3, REST API, MySQL 8, JPA,
Docker, JWT, React JS .This website is developed for a computer
institute that conducts various computer courses for all age groups. One part of the System provides information of
each course on web site and other is admin part that caters the need of back-office staff for Enquiry, Follow-up, and
Registration, Payments & Receipts.